Problem: Environment builder is currently in the `main.rs` module, but should be in the `storage` module since it’s used there.
Solution: Move the env builder into `sotarge.rs`.
