+++ src/crates.rs
@@ -46,3 +46,6 @@ extern crate crypto;
 extern crate serde_json;
 
 extern crate rand;
+
+#[macro_use]
+extern crate log;
