+++ src/main.rs
@@ -7,14 +7,12 @@
 #![cfg_attr(test, feature(test))]
 
 #![cfg_attr(not(target_os = "windows"), feature(alloc, heap_api))]
-#![cfg_attr(target_os = "windows", feature(alloc))]
+#![feature(alloc)]
 
 include!("crates.rs");
 
 extern crate num_cpus;
 extern crate log4rs;
-#[macro_use]
-extern crate log;
 
 pub mod script;
 pub mod server;
@@ -22,72 +20,25 @@ pub mod timestamp;
 pub mod pubsub;
 pub mod storage;
 
+use std::fs;
 use std::thread;
 use std::sync::Arc;
-
-use std::fs;
-#[cfg(not(target_os = "windows"))]
-use std::path;
-
-#[cfg(not(target_os = "windows"))]
-use std::ffi::CString;
-#[cfg(not(target_os = "windows"))]
-use libc::statvfs;
-
-#[cfg(not(target_os = "windows"))]
-use alloc::heap;
-#[cfg(not(target_os = "windows"))]
-use core::mem::size_of;
 
 use std::sync::Mutex;
 
 lazy_static! {
- static ref ENV: Arc<lmdb::Environment> = {
-     let _ = config::set_default("storage.path", "pumpkin.db");
 
-     let path = config::get_str("storage.path").unwrap().into_owned();
-     fs::create_dir_all(path.as_str()).expect("can't create directory");
-     unsafe {
-            let mut env_builder = lmdb::EnvBuilder::new()
-                .expect("can't create env builder");
-
-            // Configure map size
-            if !cfg!(target_os = "windows") && config::get_int("storage.mapsize").is_none() {
-                #[cfg(not(target_os = "windows"))]
-                {
-                    let path = path::PathBuf::from(path.as_str());
-                    let canonical = fs::canonicalize(&path).unwrap();
-                    let absolute_path = canonical.as_path().to_str().unwrap();
-                    let absolute_path_c = CString::new(absolute_path).unwrap();
-                    let statp: *mut statvfs = heap::allocate(size_of::<statvfs>(), size_of::<usize>()) as *mut statvfs;
-                    let mut stat = *statp;
-                    if statvfs(absolute_path_c.as_ptr(), &mut stat) != 0 {
-                       warn!("Can't determine available disk space");
-                    } else {
-                       let size = (stat.f_frsize * stat.f_bavail as u64) as usize;
-                       info!("Available disk space is approx. {}Gb, setting database map size to it", size / (1024*1024*1024));
-                       env_builder.set_mapsize(size).expect("can't set map size");
-                    }
-                    heap::deallocate(statp as *mut u8, size_of::<statvfs>(), size_of::<usize>());
-                }
-            } else {
-                match config::get_int("storage.mapsize") {
-                   Some(mapsize) => {
-                       env_builder.set_mapsize(1024 * mapsize as usize).expect("can't set map size");
-                   },
-                   None => {
-                       warn!("No default storage.mapsize set, setting it to 1Gb");
-                       env_builder.set_mapsize(1024 * 1024 * 1024).expect("can't set map size");
-                   }
-                }
-            }
-            Arc::new(env_builder
-                .open(path.as_str(), lmdb::open::Flags::empty(), 0o600)
-                .expect("can't open env"))
-    }
+ static ref ENVIRONMENT: lmdb::Environment = {
+    let _ = config::set_default("storage.path", "pumpkin.db");
+    let storage_path = config::get_str("storage.path").unwrap().into_owned();
+    fs::create_dir_all(storage_path.as_str()).expect("can't create directory");
+    let map_size = config::get_int("storage.mapsize");
+    storage::create_environment(storage_path, map_size)
  };
 
- static ref DATABASE: Arc<storage::Storage<'static>> = Arc::new(storage::Storage::new(&ENV));
+ static ref DATABASE: Arc<storage::Storage<'static>> = {
+    Arc::new(storage::Storage::new(&ENVIRONMENT))
+ };
 
  static ref PUBLISHER: Mutex<pubsub::PublisherAccessor<Vec<u8>>> = {
      let mut publisher = pubsub::Publisher::new();
